requirements.txt
streamlit==1.29.0
streamlit-authenticator==0.2.3
anthropic==0.25.0
openai==1.12.0
supabase==2.3.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
PyYAML==6.0.1
python-dotenv==1.0.0
bcrypt==4.1.2
numpy==1.26.3
